Removing backupdr_management_server_test and adding Checkdestroy to handwritten files. (#10172)
[upstream:dbf67bb7dd9fead2abfe6aeeb4cc8e28667597b8]
