+++ .changelog/10172.txt
@@ -1,0 +1,3 @@
+```release-note:none
+Fixing failing terraform integration tests
+```
