--- google-beta/services/backupdr/resource_backup_dr_management_server_generated_test.go
@@ -1,118 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-// ----------------------------------------------------------------------------
-//
-//     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-//
-// ----------------------------------------------------------------------------
-//
-//     This file is automatically generated by Magic Modules and manual
-//     changes will be clobbered when the file is regenerated.
-//
-//     Please read more about how to change this file in
-//     .github/CONTRIBUTING.md.
-//
-// ----------------------------------------------------------------------------
-
-package backupdr_test
-
-import (
-	"fmt"
-	"strings"
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
-
-	"github.com/hashicorp/terraform-provider-google-beta/google-beta/acctest"
-	"github.com/hashicorp/terraform-provider-google-beta/google-beta/envvar"
-	"github.com/hashicorp/terraform-provider-google-beta/google-beta/tpgresource"
-	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
-)
-
-func TestAccBackupDRManagementServer_backupDrManagementServerTestExample(t *testing.T) {
-	t.Parallel()
-
-	context := map[string]interface{}{
-		"project":       envvar.GetTestProjectFromEnv(),
-		"network_name":  acctest.BootstrapSharedServiceNetworkingConnection(t, "vpc-network-1"),
-		"random_suffix": acctest.RandString(t, 10),
-	}
-
-	acctest.VcrTest(t, resource.TestCase{
-		PreCheck:                 func() { acctest.AccTestPreCheck(t) },
-		ProtoV5ProviderFactories: acctest.ProtoV5ProviderBetaFactories(t),
-		CheckDestroy:             testAccCheckBackupDRManagementServerDestroyProducer(t),
-		Steps: []resource.TestStep{
-			{
-				Config: testAccBackupDRManagementServer_backupDrManagementServerTestExample(context),
-			},
-			{
-				ResourceName:            "google_backup_dr_management_server.ms-console",
-				ImportState:             true,
-				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"location", "name"},
-			},
-		},
-	})
-}
-
-func testAccBackupDRManagementServer_backupDrManagementServerTestExample(context map[string]interface{}) string {
-	return acctest.Nprintf(`
-data "google_compute_network" "default" {
-  provider = google-beta
-  name = "%{network_name}"
-}
-
-resource "google_backup_dr_management_server" "ms-console" {
-  provider = google-beta
-  location = "us-central1"
-  name     = "tf-test-ms-console%{random_suffix}"
-  type     = "BACKUP_RESTORE" 
-  networks {
-    network      = data.google_compute_network.default.id
-    peering_mode = "PRIVATE_SERVICE_ACCESS"
-  }
-}
-`, context)
-}
-
-func testAccCheckBackupDRManagementServerDestroyProducer(t *testing.T) func(s *terraform.State) error {
-	return func(s *terraform.State) error {
-		for name, rs := range s.RootModule().Resources {
-			if rs.Type != "google_backup_dr_management_server" {
-				continue
-			}
-			if strings.HasPrefix(name, "data.") {
-				continue
-			}
-
-			config := acctest.GoogleProviderConfig(t)
-
-			url, err := tpgresource.ReplaceVarsForTest(config, rs, "{{BackupDRBasePath}}projects/{{project}}/locations/{{location}}/managementServers/{{name}}")
-			if err != nil {
-				return err
-			}
-
-			billingProject := ""
-
-			if config.BillingProject != "" {
-				billingProject = config.BillingProject
-			}
-
-			_, err = transport_tpg.SendRequest(transport_tpg.SendRequestOptions{
-				Config:    config,
-				Method:    "GET",
-				Project:   billingProject,
-				RawURL:    url,
-				UserAgent: config.UserAgent,
-			})
-			if err == nil {
-				return fmt.Errorf("BackupDRManagementServer still exists at %s", url)
-			}
-		}
-
-		return nil
-	}
-}
